Refactor UserInput
