+++ shared/src/commonMain/kotlin/com/vml/tutorial/plantshop/profilePreferences/presentation/editPersonalInfo/components/EditProfileScreen.kt
@@ -14,14 +14,11 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.ArrowBack
-import androidx.compose.material3.Button
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
@@ -35,8 +32,7 @@ import com.vml.tutorial.plantshop.core.presentation.LoadingButton
 import com.vml.tutorial.plantshop.core.presentation.UiText
 import com.vml.tutorial.plantshop.core.presentation.asString
 import com.vml.tutorial.plantshop.profilePreferences.presentation.TitleSection
-import com.vml.tutorial.plantshop.profilePreferences.presentation.UserInput
-import com.vml.tutorial.plantshop.profilePreferences.presentation.editAddress.components.EditAddressEvent
+import com.vml.tutorial.plantshop.core.presentation.UserInput
 import com.vml.tutorial.plantshop.profilePreferences.presentation.editPersonalInfo.EditProfileScreenState
 import com.vml.tutorial.plantshop.ui.theme.Typography
